feat(check): Add check for non-string ipVersion
This commit introduces a new check, `IpVersionIsNotString`, to validate that the `ipVersion` property in network objects is a string as required by the specification.

The check is triggered if the `ipVersion` field, which is of type `Stringish`, is deserialized from a JSON number or boolean. A new test case has been added to ensure that this check is correctly triggered.