feat(check): Add check for non-string publicId type
This commit introduces a new check, `PublicIdTypeIsNotString`, to validate that the `type` property in publicId objects is a string as required by the specification.

The check is triggered if the `type` field, which is of type `Stringish`, is deserialized from a JSON number or boolean. A new test case has been added to ensure that this check is correctly triggered.