feat(check): Add check for non-string publicId identifier
This commit introduces a new check, `PublicIdIdentifierIsNotString`, to validate that the `identifier` property in publicId objects is a string as required by the specification.

The check is triggered if the `identifier` field, which is of type `Stringish`, is deserialized from a JSON number or boolean. A new test case has been added to ensure that this check is correctly triggered.