feat(check): Add check for non-string country code
This commit introduces a new check, `NetworkOrAutnumCountryIsNotString`, to validate that the `country` property in autnum and network objects is a string as required by the specification.

The `country` field in both `Autnum` and `Network` has been changed from `Option<String>` to `Option<Stringish>` to support this check. The builders for both structs have been updated to convert `String` to `Stringish` internally.

The check is triggered if the `country` field is deserialized from a JSON number or boolean. New test cases have been added to ensure that this check is correctly triggered for both autnum and network objects.